requirements.txt
Mako==0.9.1
MarkupSafe==0.18
PasteDeploy==1.5.2
Pygments==1.6
SQLAlchemy==0.9.3
WebOb==1.3.1
argparse==1.2.1
beautifulsoup4==4.3.2
docutils==0.11
gunicorn==18.0
mock==1.0.1
passlib==1.6.2
pyramid==1.4.5
-e git+https://github.com/Pylons/pyramid_mailer.git@778c521b3f32c9d4dc92ff1d07f8ad597a263c02#egg=pyramid_mailer
pyramid-mako==0.3.1
pyramid-tm==0.7
repoze.lru==0.6
repoze.sendmail==4.2
six==1.5.2
transaction==1.4.1
translationstring==1.1
venusian==1.0a8
waitress==0.8.8
zope.deprecation==4.1.0
zope.interface==4.1.0
zope.sqlalchemy==0.7.4
